+++ frontend/js/gens.ts
@@ -37,6 +37,7 @@ import { SampleInfo } from "./home/sample_table";
 import { setupShortcuts } from "./shortcuts";
 import { getMainSample } from "./util/utils";
 import { HelpMenu } from "./components/side_menu/help_menu";
+import { parseSex } from "./util/meta_warnings";
 
 export async function samplesListInit(
   samples: SampleInfo[],
@@ -71,6 +72,7 @@ export async function initCanvases({
   version,
   allSamples,
   defaultProfiles,
+  warningThresholds: warningThresholds,
 }: {
   caseId: string;
   sampleIds: string[];
@@ -83,6 +85,7 @@ export async function initCanvases({
   version: string;
   allSamples: Sample[];
   defaultProfiles: Record<string, ProfileSettings>;
+  warningThresholds: WarningThreshold[];
 }) {
   const gensTracks = document.getElementById("gens-tracks") as TracksManager;
   const sideMenu = document.getElementById("side-menu") as SideMenu;
@@ -141,11 +144,14 @@ export async function initCanvases({
     sampleIds.map((sampleId) => api.getSample(caseId, sampleId)),
   );
   const samples = orderSamples(unorderedSamples).map((sample) => {
+
+    const parsedSex = parseSex(sample.sex);
+
     const result: Sample = {
       caseId: sample.case_id,
       sampleId: sample.sample_id,
       sampleType: sample.sample_type,
-      sex: sample.sex,
+      sex: parsedSex,
       meta: sample.meta,
     };
     return result;
@@ -168,6 +174,7 @@ export async function initCanvases({
     api.getChromSizes(),
     startRegion,
     allAnnotSources,
+    warningThresholds,
   );
 
   const renderDataSource = getRenderDataSource(
@@ -193,43 +200,27 @@ export async function initCanvases({
     gensTracks,
   );
 
-  infoPage.setSources(() => session.getSamples());
+  infoPage.setSources(
+    () => session.getSamples(),
+    (metaId: string) => session.getMetaWarnings(metaId),
+  );
 
-  inputControls.initialize(
+  const getSearchResults = (query: string) => {
+    const annotIds = session
+      .getAnnotationSources({ selectedOnly: true })
+      .map((annot) => annot.id);
+    return api.getSearchResult(query, annotIds);
+  };
+
+  initializeInputControls(
     session,
-    async (range) => {
-      session.pos.setViewRange(range);
-      render({ reloadData: true, positionOnly: true });
-    },
-    () => {
-      sideMenu.showContent("Settings", [settingsPage], STYLE.menu.width);
-
-      if (!settingsPage.isInitialized) {
-        settingsPage.initialize();
-        render({});
-      }
-    },
-    () => {
-      sideMenu.showContent("Sample info", [infoPage], STYLE.menu.width);
-      infoPage.render();
-    },
-    () => {
-      sideMenu.showContent("Help", [helpPage], STYLE.menu.width);
-      helpPage.render();
-    },
-    () => {
-      session.toggleChromViewActive();
-      render({});
-    },
-    (query: string) => {
-      const annotIds = session
-        .getAnnotationSources({ selectedOnly: true })
-        .map((annot) => annot.id);
-      return api.getSearchResult(query, annotIds);
-    },
-    (settings: RenderSettings) => {
-      render(settings);
-    },
+    inputControls,
+    sideMenu,
+    render,
+    settingsPage,
+    infoPage,
+    helpPage,
+    getSearchResults,
   );
 
   await gensTracks.initializeTrackView(
@@ -241,6 +232,59 @@ export async function initCanvases({
   );
 
   render({ reloadData: true, resized: true });
+}
+
+function initializeInputControls(
+  session: GensSession,
+  inputControls: InputControls,
+  sideMenu: SideMenu,
+  render: (settings: RenderSettings) => void,
+  settingsPage: SettingsMenu,
+  infoPage: InfoMenu,
+  helpPage: HelpMenu,
+  getSearchResults: (query: string) => Promise<ApiSearchResult>,
+) {
+
+  const showBadge = session.hasMetaWarnings();
+
+  const onPositionChange = async (range) => {
+    session.pos.setViewRange(range);
+    render({ reloadData: true, positionOnly: true });
+  };
+  const onOpenSettings = () => {
+    sideMenu.showContent("Settings", [settingsPage], STYLE.menu.width);
+    if (!settingsPage.isInitialized) {
+      settingsPage.initialize();
+      render({});
+    }
+  };
+  const onOpenInfo = () => {
+    sideMenu.showContent("Sample info", [infoPage], STYLE.menu.width);
+    infoPage.render();
+  };
+  const onOpenHelp = () => {
+    sideMenu.showContent("Help", [helpPage], STYLE.menu.width);
+    helpPage.render();
+  };
+  const onToggleChromView = () => {
+    session.toggleChromViewActive();
+    render({});
+  };
+  const onChange = (settings: RenderSettings) => {
+    render(settings);
+  };
+
+  inputControls.initialize(
+    session,
+    onPositionChange,
+    onOpenSettings,
+    onOpenInfo,
+    onOpenHelp,
+    onToggleChromView,
+    getSearchResults,
+    onChange,
+    showBadge,
+  );
 }
 
 function addSettingsPageSources(